Refine stats query semantics and tests
